+++ src/gestorAplicacion/inferior/ContactosLocales.java
@@ -1,25 +1,31 @@
-package gestorAplicacion.hija;
+package gestorAplicacion.inferior;
 
-import gestorAplicacion.padre.Contacto;
+import java.util.ArrayList;
 
-//Los contactos pendientes son personas con las que no se a iniciado un conversacion.
-class ContactosPendientes extends Contacto{
-	private final String mensajes;
+import gestorAplicacion.superior.Contacto;
+import gestorAplicacion.superior.Mensaje;
 
-	//Crea el contacto sin mensajes
-	ContactosPendientes(String nombre, String email, String nombreCompleto) {
+//Los contactos locales son personas con las que ya se inicializo una convesacion
+class ContactosLocales extends Contacto{
+	private ArrayList<Mensaje> mensajes;
+
+	//Genera los contactos locales con la capacidad de resiver mensajes
+	ContactosLocales(String nombre, String email, String nombreCompleto) {
 		super(nombre, email, nombreCompleto);
-		this.mensajes = "No hay mensajes";
+		mensajes = new ArrayList<Mensaje>();
 	}
 
-	String getTargeta() {
+	ArrayList<Mensaje> getMensajes() {
 		return mensajes;
 	}
-	
+
+	void setMensajes(ArrayList<Mensaje> mensajes) {
+		this.mensajes = mensajes;
+	}
+
 	@Override
-	//FIXME
 	public String toString() {
-		return super.toString() + "\nContactosPendientes [mensajes=" + mensajes + "]";
+		return super.toString();
 	}
 
 	@Override
