+++ src/gestorAplicacion/superior/Mensaje.java
@@ -1,4 +1,4 @@
-package gestorAplicacion.padre;
+package gestorAplicacion.superior;
 
 import java.time.LocalDateTime;
